pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Konstantin Kiselev <konstantin.kiselev.1998@mail.ru>"]
readme = "README.md"
packages = [
    { include = "telegram_bestbet_bot" },
]
[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.10.4"
aiologger = "^0.7.0"
requests = "^2.32.3"
aiogram = "^3.12.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
